refactors sync activity into independent module
